Extract IncompatibleTransactionsVerifier
