+++ packages/processor/source/verifiers/index.ts
@@ -1,1 +1,2 @@
+export * from "./incompatible-transactions-verifier";
 export * from "./verify-block-verifier";
